+++ site/admin/usuario/UsuarioList.php
@@ -2,11 +2,11 @@
 include "../header.php";
 include "../db.class.php";
 
-$db = new db('usuario', 'ids');
+$db = new db('usuario', 'idusuarios');
 
 $db->checkLogin();
 
-if (!empty($_GET['id'])) {
+if (!empty($_GET['id']) && !empty($_GET['action']) && $_GET['action'] === 'delete') {
     $db->destroy($_GET['id']);
     header('Location: UsuarioList.php');
     exit;
@@ -26,7 +26,7 @@ if (!empty($_POST)) {
 <form action="./UsuarioList.php" method="post">
     <div class="row">
         <div class="col">
-            <select name="usuario" class="form-select">
+            <select name="tipo" class="form-select">
                 <option value="nome">nome</option>
                 <option value="telefone">Telefone</option>
                 <option value="email">Email</option>
@@ -64,14 +64,14 @@ if (!empty($_POST)) {
                     if($dados) {
                         foreach ($dados as $item) {
                             echo "<tr>
-                                <th scope='row'>$item->ids</th>
+                                <th scope='row'>$item->idusuarios</th>
                                 <td>$item->nome</td>
                                 <td>$item->telefone</td>
                                 <td>$item->email</td>
                                 <td>$item->login</td>     
                                 <td>
-                                    <a href='./UsuarioForm.php?id=$item->ids' class='btn btn-warning btn-sm'>Editar</a>
-                                    <a href='./UsuarioList.php?id=$item->ids'
+                                    <a href='./UsuarioForm.php?id=$item->idusuarios' class='btn btn-warning btn-sm'>Editar</a>
+                                    <a href='./UsuarioList.php?action=delete&id=$item->idusuarios'
                                        onclick='return confirm(\"Deseja realmente excluir?\")'
                                        class='btn btn-danger btn-sm'>
                                        Excluir
